fix(backend): [PYTA-63] move stocks.py to correct dir
