+++ backend/app/v1/auth/__init__.py
@@ -1,4 +1,4 @@
 from app.v1.auth.auth_router import auth_router
-from app.v1.auth.stocks import stocks_router
+from app.v1.stocks.stocks import stocks_router
 
 __all__ = ["auth_router", "stocks_router"]
