+++ backend/app/v1/stocks/stocks.py
